==> 3.chatgpt.ipynb <==
# This Python 3 environment comes with many helpful analytics libraries installed
# It is defined by the kaggle/python Docker image: https://github.com/kaggle/docker-python
# For example, here's several helpful packages to load

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)

# Input data files are available in the read-only "../input/" directory
# For example, running this (by clicking run or pressing Shift+Enter) will list all files under the input directory

import os
for dirname, _, filenames in os.walk('/kaggle/input'):
    for filename in filenames:
        print(os.path.join(dirname, filename))

# You can write up to 20GB to the current directory (/kaggle/working/) that gets preserved as output when you create a version using "Save & Run All" 
# You can also write temporary files to /kaggle/temp/, but they won't be saved outside of the current session

df = pd.read_csv('/kaggle/input/nlpsum2/sum_data (1).csv')

df

!pip install transformers 

from transformers import pipeline

summarizer = pipeline('summarization')

summarizer(df.loc[0]['content'],min_length = 30,max_length = 130,do_sample = False)

df.loc[0]['content']

import torch
import json 
from transformers import T5Tokenizer, T5ForConditionalGeneration, T5Config

model = T5ForConditionalGeneration.from_pretrained('t5-small')
tokenizer = T5Tokenizer.from_pretrained('t5-small')
device = torch.device('cpu')

text = df.loc[1]['content']


preprocess_text = text.strip().replace("\n","")
t5_prepared_Text = "summarize: "+preprocess_text
print ("original text preprocessed: \n", preprocess_text)

tokenized_text = tokenizer.encode(t5_prepared_Text, return_tensors="pt").to(device)


# summmarize 
summary_ids = model.generate(tokenized_text,
                                    num_beams=4,
                                    no_repeat_ngram_size=2,
                                    min_length=30,
                                    max_length=100,
                                    early_stopping=True)

output = tokenizer.decode(summary_ids[0], skip_special_tokens=True)

print ("\n\nSummarized text: \n",output)



!pip install torch==1.8.2+cu111 torchvision==0.9.2+cu111 torchaudio===0.8.2 -f https://download.pytorch.org/whl/lts/1.8/torch_lts.html

!pip install transformers
!pip install sentencepiece

pip install ipywidgets

df['Summary'] = ''

c = 0
for ind in range(len(df)):
    print(df['content'][ind])
    c  = c+1
    if c == 5:
        break

from bs4 import BeautifulSoup
from tqdm import tqdm
import torch
device = 'cuda' if torch.cuda.is_available() else 'cpu'

from transformers import PegasusForConditionalGeneration, PegasusTokenizer
tokenizer = PegasusTokenizer.from_pretrained("google/pegasus-xsum")
model = PegasusForConditionalGeneration.from_pretrained("google/pegasus-xsum").to(device)

device = torch.device("cuda")
model.cuda()

l=[]

for ind in tqdm(range(len(df)),desc="Loading..."):
    tokens = tokenizer(df['content'][ind], truncation=True, padding="longest", return_tensors="pt")
    summary = model.generate(**tokens.to(device),no_repeat_ngram_size = 1,min_new_tokens = 300)
    l.append(BeautifulSoup(tokenizer.decode(summary[0]), 'lxml').get_text())

df['summary'] = l

df.to_csv("f_peg_sum.csv")

df[546:]

Using Chatgpt to Generate Summary


pip install openai

import openai

openai.api_key='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'

r = []
for ind in tqdm(range(len(df)),desc="Loading..."):
    prompt = "Summarize this set of tweets:"+df['content'][ind]
    completion = openai.Completion.create(
    engine = "text-davinci-003",
    prompt=prompt,
    max_tokens=300,
    n=1,
    stop=None,
    temperature=0.5)
    r.append(completion.choices[0].text)

len(r)

r[545]

df = df[546:]

df

for ind in tqdm(range(len(df)),desc="Loading..."):
    prompt = "Summarize this set of tweets:"+df['content'][ind+546]
    completion = openai.Completion.create(
    engine = "text-davinci-003",
    prompt=prompt,
    max_tokens=300,
    n=1,
    stop=None,
    temperature=0.5)
    r.append(completion.choices[0].text)

r[0]

tweet_df = pd.DataFrame(r,columns=["c_r_sum"])

tweet_df

tweet_df.to_csv("sum007.csv")

len(z)

2*[0]

len(r+454*[''])

test =  pd.read_csv('/kaggle/working/sum007.csv')

test
