==> 4.Summary_Generation_For_Dataset_Using_Classical_Approach.ipynb <==
#RANKING THE SENTENCES USING POS-TAGGING and TAKING THE TOP RANKED SENTENCES

import nltk
nltk.download("stopwords")
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')

import math
from nltk.corpus import stopwords
from nltk.tokenize import sent_tokenize, word_tokenize
from nltk.tag import pos_tag
from collections import defaultdict
text = """
The quick brown fox jumped over the lazy dog. The dog barked loudly. The fox was startled and ran away. 
The cat watched from the tree. The sky turned dark and it started to rain heavily.
"""

sentences = sent_tokenize(text)
tagged_sentences = [pos_tag(word_tokenize(sent)) for sent in sentences]
tagged_sentences

  sentence_scores = defaultdict(int)
  sentence_scores

stop_words = set(stopwords.words("english"))
important_pos_tags = ["NN", "NNS", "VB", "VBD", "VBG", "VBN", "VBP", "VBZ", "JJ", "JJR", "JJS"]
for i in range(len(tagged_sentences)):
        for word, pos_tag in tagged_sentences[i]:
            if pos_tag in important_pos_tags and word.lower() not in stop_words:
                sentence_scores[i] += 1

sentence_scores

sorted_sentences = sorted(sentence_scores, key=sentence_scores.get, reverse=True)
sorted_sentences

candidate_sentences = sorted_sentences[:3]
sentence_scores
summary = " ".join([sentences[i] for i in sorted(candidate_sentences)])
summary

MAKING FUNCTION FOR POS TAGGING

import nltk
import math
from nltk.corpus import stopwords
from nltk.tokenize import sent_tokenize, word_tokenize
from nltk.tag import pos_tag
from collections import defaultdict

# Load the NLTK stop words
stop_words = set(stopwords.words("english"))

# Example usage
text = """
The quick brown fox jumped over the lazy dog. The dog barked loudly. The fox was startled and ran away. 
The cat watched from the tree. The sky turned dark and it started to rain heavily.
"""

def pos_tagging_summarization(text, num_sentences=math.ceil(len(sent_tokenize(text))*.5)):
    # Tokenize the text into sentences
    sentences = sent_tokenize(text)
    
    # Tokenize the sentences into words and perform POS tagging
    tagged_sentences = [pos_tag(word_tokenize(sent)) for sent in sentences]
    
    # Create a dictionary to store the sentence scores
    sentence_scores = defaultdict(int)
    
    # Define the POS tags that are considered important
    important_pos_tags = ["NN", "NNS", "VB", "VBD", "VBG", "VBN", "VBP", "VBZ", "JJ", "JJR", "JJS"]
    
    # Loop through each tagged sentence and calculate sentence scores based on the occurrence of important POS tags
    for i in range(len(tagged_sentences)):
        for word, p_t in tagged_sentences[i]:
            if p_t in important_pos_tags and word.lower() not in stop_words:
                sentence_scores[i] += 1
    
    # Sort the sentences based on their scores in descending order
    sorted_sentences = sorted(sentence_scores, key=sentence_scores.get, reverse=True)
    
    # Select the top 'num_sentences' sentences as candidate summary sentences
    candidate_sentences = sorted_sentences[:num_sentences]
    
    # Sort the candidate sentences based on their occurrence in the original text
    summary = " ".join([sentences[i] for i in sorted(candidate_sentences)])
    
    return summary


summary = pos_tagging_summarization(text)
print("Summary:")
print(summary)

#GENERATING DATA USING MARKOV MODEL

import numpy as np
import pandas as pd
import os
import re
import string
from nltk.tokenize import word_tokenize,sent_tokenize
from nltk.corpus import stopwords
import random

text = """
The quick brown fox jumped over the lazy dog. The dog barked loudly. The fox was startled and ran away. 
The cat watched from the tree. The sky turned dark and it started to rain heavily.
"""
def clean_txt(txt):
    cleaned_txt = []
    sentences = sent_tokenize(text)

    for line in sentences:
        line = line.lower()
        line = re.sub(r"[,.\"\'!@#$%^&*(){}?/;`~:<>+=-\\]", "", line)
        tokens = word_tokenize(line)
        words = [word for word in tokens if word.isalpha()]
        cleaned_txt+=words
    return cleaned_txt

cleaned_text = clean_txt(text)
print("number of words = ", len(cleaned_text))

cleaned_text

tags = pos_tag(cleaned_text)
tags

stop_words = set(stopwords.words("english"))
tagged_cleaned_text = []
important_pos_tags = ["NN", "NNS", "VB", "VBD", "VBG", "VBN", "VBP", "VBZ", "JJ", "JJR", "JJS"]
for i in tags:
  if i[1] in important_pos_tags:
    tagged_cleaned_text.append(i[0])
len(tagged_cleaned_text)


def make_markov_model(cleaned_stories, n_gram=2):    #to get contextual info we are using n_gram
    markov_model = {}
    for i in range(len(cleaned_stories)-n_gram-1):
        curr_state, next_state = "", ""
        for j in range(n_gram):
            curr_state += cleaned_stories[i+j] + " "
            next_state += cleaned_stories[i+j+n_gram] + " "
        curr_state = curr_state[:-1]
        next_state = next_state[:-1]
        
        if curr_state not in markov_model:
            markov_model[curr_state] = {}
            markov_model[curr_state][next_state] = 1
        else:
            if next_state in markov_model[curr_state]:
                markov_model[curr_state][next_state] += 1
            else:
                markov_model[curr_state][next_state] = 1
    # calculating transition probabilities
    for curr_state, transition in markov_model.items():
        total = sum(transition.values())
        for state, count in transition.items():
            markov_model[curr_state][state] = count/total
    return markov_model

markov_model = make_markov_model(tagged_cleaned_text,2)
markov_model

print("number of states = ", len(markov_model.keys()))

def generate_story(markov_model, limit=10, start=''):
    n = 0
    curr_state = start
    next_state = None
    story = ""
    story+=curr_state+" "
    while n<limit:
      try:
        next_state = random.choices(list(markov_model[curr_state].keys()),
                                    list(markov_model[curr_state].values()))
        curr_state = next_state[0]
        story+=curr_state+" "
        n+=1
      except KeyError:
        return story
    return story

list(markov_model.keys())

stories = []
for i in range(5):
  import random


  # Get a random key from the dictionary
  random_key = random.choice(list(markov_model.keys()))
  print(random_key)
  story = generate_story(markov_model,10,start=str(random_key))
  stories.append(story)
  


stories

stories

**USING TEXT RANK ON STORIES To Generate Summary**

import nltk
from nltk.tokenize import sent_tokenize, word_tokenize
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from string import punctuation
from collections import defaultdict
nltk.download('wordnet')



!wget http://nlp.stanford.edu/data/glove.6B.zip

!unzip glove*.zip
     

import numpy as np
#Extract word vectors
word_embeddings = {}
f = open('glove.6B.100d.txt',encoding = 'utf-8')
for line in f:
  values = line.split()
  word = values[0]
  coefs = np.asarray(values[1:],dtype = 'float32')
  word_embeddings[word] = coefs
f.close()

nltk.download('omw-1.4')
nltk.download('stopwords')
stop_words = set(stopwords.words('english') + list(punctuation))
lemmatizer = WordNetLemmatizer()

def preprocess(sentence):
    words = word_tokenize(sentence.lower())
    words = [lemmatizer.lemmatize(word) for word in words if word not in stop_words]
    return words

processed_sentences = [preprocess(sentence) for sentence in stories]

processed_sentences

#vectorize sentences
sentence_vectors = []
for i in processed_sentences:
  if len(i) != 0:
    v = sum([word_embeddings.get(w,np.zeros((100,))) for w in i])/(len(i)+0.001)
  else:
    v = np.zeros((100,))
  sentence_vectors.append(v)

len(sentence_vectors)

# similarity matrix
sim_mat = np.zeros([len(stories),len(stories)])

from sklearn.metrics.pairwise import cosine_similarity
     

for i in range(len(stories)):
  for j in range(len(stories)):
    if i!=j:
      sim_mat[i][j] = cosine_similarity(sentence_vectors[i].reshape(1,100),sentence_vectors[j].reshape(1,100))[0,0]

     

!pip install scipy==1.8.0 
!pip install networkx==2.6
     

#Creating a graph
import networkx as nx
import matplotlib.pyplot as plt
     

nx_graph = nx.from_numpy_array(sim_mat)
scores = nx.pagerank(nx_graph)
scores
     

ranked_sentences = sorted(((scores[i],s) for i,s in enumerate(stories)),reverse=True)
     

ranked_sentences

import math
s=""
for i in range(math.ceil(len(stories)*(.5))):
  s=s+ranked_sentences[i][1]

print(s)
