==> 5.Final dataset creation.ipynb <==
import pandas as pd

df = pd.read_csv('sum_data (1).csv')
df1 = pd.read_csv('sum007.csv')
df2 = pd.read_csv('sum008.csv')
df3 = pd.read_csv('sum009.csv')

type(df2['c_r_sum'])

l1 = df1['c_r_sum'].to_list()
l2 = df2['c_r_sum'].to_list()
l3 = df3['c_r_sum'].to_list()

l = l1+l2+l3

len(l)

c = df['content'].to_list()

len(c)

c = c[0:983]

len(c)

v = df['topic'].to_list()
v = v[0:983]

f = []
for i in range(983):
    t = [v[i],c[i],l[i]]
    f.append(t)

len(f)

final_df = pd.DataFrame(f,columns=["topic","content",'c_summary'])

final_df

def removen(x):
    return x.replace("\n","")
final_df['c_summary'] = final_df['c_summary'].apply(lambda x: removen(x) )

final_df

final_df.to_csv("final_data.csv")

out_df = pd.read_csv('tweetsum_train.csv')
in_df = pd.read_csv('final_data.csv')
val_df = pd.read_csv('tweetsum_valid.csv')

out_df

in_df

val_df

l1 = out_df['inputs'].to_list()
l2 = in_df['content'].to_list()
l3 = val_df['inputs'].to_list()

l = l2+l1+l3

len(l)

s1 = out_df['summaries'].to_list()
s2 = in_df['c_summary'].to_list()
s3 = val_df['summaries'].to_list()

s = s2+s1+s3

len(s)

f = []
for i in range(1960):
    t = [l[i],s[i]]
    f.append(t)

f_df = pd.DataFrame(f,columns=["content",'c_summary'])

f_df

f_df.to_csv("NLP.csv")
